// apps/web/src/lib/api/generated/models/bridgesUpdateBody.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Formex
 * OpenAPI spec version: 0.0.1
 */

export type BridgesUpdateBody = {
  curriculum_gap: string;
  details: string;
  conducted_date: string;
  venue: string;
  resource_person_name: string;
  company_name: string;
  designation: string;
  students_present: number;
  relevance: string;
  academic_year: number;
  course_coordinator_id: string;
};